Fixed escaping of path
Related to this [issue](https://github.com/ngCli/ng-cli/issues/6). I've tested my addition in Windows 8.1 and it is working now.
The problem was that the generated js object in snippet.js was not valid because the \ was not properly escaped. Before my code the snippet had the following keys 'content\controller\controller.coffee' and that's not valid for json and now it is like - 'content\\controller\\controller.coffee' with escaped back-slashes. The error message mentioned in the issue disapeared.